bot-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
pydantic==2.5.0
python-dotenv==1.0.0
langchain==0.0.350
langchain-community==0.0.10
python-multipart==0.0.6
httpx==0.25.2 